+++ CHANGES.md
@@ -44,6 +44,7 @@ Beta Releases
    * Removed `canvasDimensions` from `FrameState`.
    * Removed `context` option from `Material` constructor and parameter from `Material.fromType`.
    * Renamed `TextureWrap.CLAMP` to `TextureWrap.CLAMP_TO_EDGE`.
+   * All data sources must now implement `DataSource.getName`, which returns a user-readable name for the data source.
 * Added `CorridorOutlineGeometry`.
 * Added `PolylineGeometry`, `PolylineColorAppearance`, and `PolylineMaterialAppearance`.
 * Added `colors` option to `SimplePolylineGeometry` for per vertex or per segment colors.
@@ -73,6 +74,7 @@ Beta Releases
   becomes
 
         return primitives.add(new Primitive(/* ... */));
+* Added `DynamicObject.name` property, which is a non-unique, user-readable name for the object.
 
 ### b20 - 2013-09-03